+++ dev/devcheck_downloader.py
@@ -7,7 +7,7 @@ def check_fsspec():
     import ubelt as ub
     from os.path import join
     dpath = ub.ensure_app_cache_dir('ubelt/simple_server')
-    info = ub.cmd(['python', '-m', 'http.server', '--directory', dpath], detatch=True)
+    info = ub.cmd(['python', '-m', 'http.server', '--directory', dpath], detach=True)
 
     fnames = ['file_{}.txt'.format(i) for i in range(100)]
     for fname in fnames: